FIX: Fix tests

+++ vispy/testing/_runners.py
@@ -14,64 +14,75 @@ from functools import partial
 
 from ..util import use_log_level, run_subprocess
 from ..util.ptime import time
-from ._testing import SkipTest, has_backend, has_application, nottest
+from ._testing import SkipTest, has_application, nottest
 
 
 _line_sep = '-' * 70
 
 
-def _get_root_dir():
-    root_dir = os.getcwd()
-    if (op.isfile(op.join(root_dir, 'setup.py')) and
-            op.isdir(op.join(root_dir, 'vispy'))):
+def _get_import_dir():
+    import_dir = op.abspath(op.join(op.dirname(__file__), '..'))
+    up_dir = op.join(import_dir, '..')
+    if (op.isfile(op.join(up_dir, 'setup.py')) and
+            op.isdir(op.join(up_dir, 'vispy')) and
+            op.isdir(op.join(up_dir, 'examples'))):
         dev = True
     else:
-        root_dir = op.abspath(op.join(op.dirname(__file__), '..', '..'))
-        dev = True if op.isfile(op.join(root_dir, 'setup.py')) else False
-    return root_dir, dev
+        dev = False
+    return import_dir, dev
 
 
 _unit_script = """
-import pytest
+try:
+    import pytest as tester
+except ImportError:
+    import nose as tester
 try:
     import faulthandler
     faulthandler.enable()
 except Exception:
     pass
 
-raise SystemExit(pytest.main(%r))
+raise SystemExit(tester.main(%r))
 """
 
 
-def _unit(mode, extra_arg_string):
+def _unit(mode, extra_arg_string, coverage=False):
     """Run unit tests using a particular mode"""
-    cwd = os.getcwd()
+    import_dir = _get_import_dir()[0]
+    cwd = op.join(import_dir, '..')
+    extra_args = [''] + extra_arg_string.split(' ')
+    del extra_arg_string
+    use_pytest = False
     try:
         import pytest  # noqa, analysis:ignore
+        use_pytest = True
     except ImportError:
-        print('Skipping pytest, pytest not installed')
-        raise SkipTest()
+        try:
+            import nose  # noqa, analysis:ignore
+        except ImportError:
+            raise SkipTest('Skipping unit tests, pytest not installed')
 
     if mode == 'nobackend':
         msg = 'Running tests with no backend'
-        extra_arg_string = '-m "not vispy_app_test" ' + extra_arg_string
-        coverage = True
+        if use_pytest:
+            extra_args += ['-m', '"not vispy_app_test"']
+        else:
+            extra_args += ['-a', '"!vispy_app_test"']
     else:
-        with use_log_level('warning', print_msg=False):
-            has, why_not = has_backend(mode, out=['why_not'])
-        if not has:
-            msg = ('Skipping tests for backend %s, not found (%s)'
-                   % (mode, why_not))
-            print(_line_sep + '\n' + msg + '\n' + _line_sep + '\n')
-            raise SkipTest(msg)
         msg = 'Running tests with %s backend' % mode
-        extra_arg_string = '-m vispy_app_test ' + extra_arg_string
-        coverage = True
-    if coverage:
-        extra_arg_string += ' --cov vispy --no-cov-on-fail '
+        if use_pytest:
+            extra_args += ['-m', 'vispy_app_test']
+        else:
+            extra_args += ['-a', 'vispy_app_test']
+    if coverage and use_pytest:
+        extra_args += ['--cov', 'vispy', '--no-cov-on-fail']
     # make a call to "python" so that it inherits whatever the system
     # thinks is "python" (e.g., virtualenvs)
-    cmd = [sys.executable, '-c', _unit_script % extra_arg_string]
+    extra_arg_string = ' '.join(extra_args)
+    insert = extra_arg_string if use_pytest else extra_args
+    extra_args += [import_dir]  # positional argument
+    cmd = [sys.executable, '-c', _unit_script % insert]
     env = deepcopy(os.environ)
 
     # We want to set this for all app backends plus "nobackend" to
@@ -83,11 +94,11 @@ def _unit(mode, extra_arg_string):
                % (_line_sep, msg, env_str, extra_arg_string))
         print(msg)
     sys.stdout.flush()
-    return_code = run_subprocess(cmd, return_code=True, cwd=cwd, env=env,
-                                 stdout=None, stderr=None)[2]
+    return_code = run_subprocess(cmd, return_code=True, cwd=cwd,
+                                 env=env, stdout=None, stderr=None)[2]
     if return_code:
         raise RuntimeError('unit failure (%s)' % return_code)
-    else:
+    if coverage:
         out_name = '.coverage.%s' % mode
         if op.isfile(out_name):
             os.remove(out_name)
@@ -97,12 +108,12 @@ def _unit(mode, extra_arg_string):
 def _flake():
     """Test flake8"""
     orig_dir = os.getcwd()
-    root_dir, dev = _get_root_dir()
-    os.chdir(root_dir)
+    import_dir, dev = _get_import_dir()
+    os.chdir(op.join(import_dir, '..'))
     if dev:
         sys.argv[1:] = ['vispy', 'examples', 'make']
     else:
-        sys.argv[1:] = ['vispy']
+        sys.argv[1:] = [op.basename(import_dir)]
     sys.argv.append('--ignore=E226,E241,E265,E266,W291,W293,W503')
     sys.argv.append('--exclude=six.py,py24_ordereddict.py,glfw.py,'
                     '_proxy.py,_es2.py,_gl2.py,_pyopengl2.py,'
@@ -135,16 +146,14 @@ def _check_line_endings():
     print('Running line endings check... ')
     sys.stdout.flush()
     report = []
-    root_dir, dev = _get_root_dir()
-    if not dev:
-        root_dir = op.join(root_dir, 'vispy')
-    for dirpath, dirnames, filenames in os.walk(root_dir):
+    import_dir, dev = _get_import_dir()
+    for dirpath, dirnames, filenames in os.walk(import_dir):
         for fname in filenames:
             if op.splitext(fname)[1] in ('.pyc', '.pyo', '.so', '.dll'):
                 continue
             # Get filename
             filename = op.join(dirpath, fname)
-            relfilename = op.relpath(filename, root_dir)
+            relfilename = op.relpath(filename, import_dir)
             # Open and check
             try:
                 with open(filename, 'rb') as fid:
@@ -203,7 +212,7 @@ def _examples(fnames_str):
         Can be a space-separated list of paths to test, or an empty string to
         auto-detect and run all examples.
     """
-    root_dir, dev = _get_root_dir()
+    import_dir, dev = _get_import_dir()
     reason = None
     if not dev:
         reason = 'Cannot test examples unless in vispy git directory'
@@ -226,7 +235,7 @@ def _examples(fnames_str):
 
     else:
         fnames = [op.join(d[0], fname)
-                  for d in os.walk(op.join(root_dir, 'examples'))
+                  for d in os.walk(op.join(import_dir, '..', 'examples'))
                   for fname in d[2] if fname.endswith('.py')]
 
     fnames = sorted(fnames, key=lambda x: x.lower())
@@ -282,7 +291,7 @@ def _examples(fnames_str):
 
 
 @nottest
-def test(label='full', extra_arg_string=''):
+def test(label='full', extra_arg_string='', coverage=False):
     """Test vispy software
 
     Parameters
@@ -292,6 +301,8 @@ def test(label='full', extra_arg_string=''):
         'flake', or any backend name (e.g., 'qt').
     extra_arg_string : str
         Extra arguments to sent to ``pytest``.
+    coverage : bool
+        If True, collect coverage data.
     """
     from vispy.app.backends import BACKEND_NAMES as backend_names
     label = label.lower()
@@ -302,16 +313,18 @@ def test(label='full', extra_arg_string=''):
     if label not in known_types + backend_names:
         raise ValueError('label must be one of %s, or a backend name %s, '
                          'not \'%s\'' % (known_types, backend_names, label))
-    work_dir = _get_root_dir()[0]
-    orig_dir = os.getcwd()
     # figure out what we actually need to run
     runs = []
     if label in ('full', 'unit'):
         for backend in backend_names:
-            runs.append([partial(_unit, backend, extra_arg_string),
+            runs.append([partial(_unit, backend, extra_arg_string, coverage),
                          backend])
     elif label in backend_names:
-        runs.append([partial(_unit, label, extra_arg_string), label])
+        runs.append([partial(_unit, label, extra_arg_string, coverage), label])
+
+    if label in ('full', 'unit', 'nobackend'):
+        runs.append([partial(_unit, 'nobackend', extra_arg_string, coverage),
+                     'nobackend'])
 
     if label == "examples":
         # take the extra arguments so that specific examples can be run
@@ -321,9 +334,6 @@ def test(label='full', extra_arg_string=''):
         # run all the examples
         runs.append([partial(_examples, ""), 'examples'])
 
-    if label in ('full', 'unit', 'nobackend'):
-        runs.append([partial(_unit, 'nobackend', extra_arg_string),
-                     'nobackend'])
     if label in ('full', 'extra', 'lineendings'):
         runs.append([_check_line_endings, 'lineendings'])
     if label in ('full', 'extra', 'flake'):
@@ -333,7 +343,6 @@ def test(label='full', extra_arg_string=''):
     skip = []
     for run in runs:
         try:
-            os.chdir(work_dir)
             run[0]()
         except RuntimeError as exp:
             print('Failed: %s' % str(exp))
@@ -351,7 +360,6 @@ def test(label='full', extra_arg_string=''):
             print('Passed\n')
         finally:
             sys.stdout.flush()
-            os.chdir(orig_dir)
     dt = time() - t0
     stat = '%s failed, %s skipped' % (fail if fail else 0, skip if skip else 0)
     extra = 'failed' if fail else 'succeeded'
